Update typedef ee_f32 from double to float.

+++ barebones/core_portme.h
@@ -87,7 +87,7 @@ Original Author: Shay Gal-on
 typedef signed short   ee_s16;
 typedef unsigned short ee_u16;
 typedef signed int     ee_s32;
-typedef double         ee_f32;
+typedef float          ee_f32;
 typedef unsigned char  ee_u8;
 typedef unsigned int   ee_u32;
 typedef ee_u32         ee_ptr_int;
